started api

+++ gui/GUI.py
@@ -1,22 +1,22 @@
 import sys 
 import wx  
-import FileDB
-import syncmonster-desktop.File
+import APIReciever
+import pynotify
 
 class Mywin(wx.Frame): 
             
     def __init__(self, parent, title): 
-        super(Mywin, self).__init__(parent, title = title, size =(600,300)) 
+        super(Mywin, self).__init__(parent, title = title, size =(400,300)) 
                 
         panel = wx.Panel(self) 
         box = wx.BoxSizer(wx.VERTICAL)
                 
         self.lst = wx.ListCtrl(panel, -1, style = wx.LC_REPORT) 
-        self.lst.InsertColumn(0, 'File Name', width = 150) 
+        self.lst.InsertColumn(0, 'File Name', width = 200) 
         self.lst.InsertColumn(1, 'size', wx.LIST_FORMAT_RIGHT, 100) 
         self.lst.InsertColumn(2, 'parts', wx.LIST_FORMAT_RIGHT, 100) 
         self.lst.InsertColumn(3, 'file id', wx.LIST_FORMAT_RIGHT, 0)
-       
+        
         self.update_files()
 
         self.upload_button = wx.Button(panel, label='Upload')
@@ -42,13 +42,17 @@ class Mywin(wx.Frame):
 
     def update_files (self):
         self.lst.DeleteAllItems()
-        files = FileDB.update()
-        print 'Updating Files...'
-        for f in files: #same as above
-            index = self.lst.InsertStringItem(sys.maxint, f[1]) # file name
-            self.lst.SetStringItem(index, 1, str(f[2])) # add later
-            self.lst.SetStringItem(index, 2, str(f[2])) #sets parts
-            self.lst.SetStringItem(index, 3, str(f[0])) # file id
+        try:
+            files = APIReciever.getFiles()
+                
+            print 'Updating Files...'
+            for f in files: #same as above
+                index = self.lst.InsertStringItem(sys.maxint, str(f[1])) # file name
+                self.lst.SetStringItem(index, 1, str(f[2])) # add later
+                self.lst.SetStringItem(index, 2, str(f[2])) #sets parts
+                self.lst.SetStringItem(index, 3, str(f[0])) # file id
+        except:
+            pass
         wx.CallLater(5000, self.update_files)
 
     def openFile(self, e):
@@ -57,18 +61,35 @@ class Mywin(wx.Frame):
         file_path = fileDialog.GetPath()
         fileDialog.Destroy()
 
+        multidlg  = wx.MultiChoiceDialog(self,
+                                        'File Settings',
+                                        'Upload',
+                                        ['Compress File', 'Split file based on size ratio of storage'])
+        multidlg.ShowModal()
+        print multidlg.GetSelections() 
+        self.showNotify('Uploaded File')
+
+
     def deleteButton (self, e):
         item = self.lst.GetFocusedItem()
         file_id = self.lst.GetItem(item, 3).GetText()
         if file_id:
-            FileDB.delete(file_id)
-            self.lst.DeleteItem(self.lst.GetFocusedItem())
+            name = self.lst.GetItemText(self.lst.GetFocusedItem())
+            dlg = wx.MessageBox("Would you like to delete " + name + '?', 'Delete', wx.YES_NO | wx.NO_DEFAULT)
+            if dlg == 2:
+                #make sure to add authentication
+                APIReciever.deleteFile(file_id)
+                self.lst.DeleteItem(self.lst.GetFocusedItem())
 
     def downloadButton (self, e):
         print self.lst.GetItemText(self.lst.GetFocusedItem())
-  
-FileDB = file_checker.FileDB()
-files = FileDB.update()
+        self.showNotify('Finished Download')
+
+    def showNotify (self, message):
+        pynotify.init('Message')
+        notice = pynotify.Notification('SyncMonster', message)
+        notice.show()
+
 ex = wx.App() 
 frame = Mywin(None,'SyncMonster') 
 frame.Show()
